change 'ollama-server' dir to 'flask' for clarity
