+++ flask/scripts/create_index_stores.py
