--- ollama-server/blueprints/__init__.py
@@ -1,2 +1,0 @@
-from .ai_assistant_bp import ai_bp
-from .filesystem_bp import fs_bp
