--- ollama-server/config/config.conf
@@ -1,14 +1,0 @@
-[paths]
-INDEX_BIN_PATH = index/faiss_index.bin     
-METADATA_DB_PATH = index/file_metadata.db   
-
-[flask]
-FLASK_PORT = 8321    
-
-[index]
-EMBEDDING_DIM = 384  
-K = 3                
-
-[ollama]
-OLLAMA_URL = http://localhost:11434
-OLLAMA_MODEL = llama3.2:3b-instruct-fp16
