--- ollama-server/utils/__init__.py
@@ -1,6 +1,0 @@
-from .general import *
-from .indexer_utils import *
-from .metadata_extraction_utils import *
-from .model_utils import *
-from .nlp_utils import *
-from .text_extraction_utils import *
